+++ go-app/frontend/src/tabs/main/MainTabProfileSelector.tsx
@@ -1,0 +1,137 @@
+import { Controller, UseFormReturn, useWatch } from "react-hook-form";
+import { main } from "../../../wailsjs/go/models";
+import { useCallback, useMemo } from "react";
+import { ProfileSelectionListItem } from "./_components/ProfileSelectionListItem";
+import { unfocus } from "../../utils/unfocus";
+import {
+  ProfileInfo,
+  ProfileSelectionMoreMenu,
+} from "./_components/ProfileSelectionMoreMenu";
+
+type Props = {
+  form: UseFormReturn<{
+    profiles: Partial<Record<string, main.Interop_SelectedProfileInfo>>;
+  }>;
+  controllers: main.Interop_GenericController[];
+  profiles: main.Interop_Profile[];
+  onReloadConfiguration: () => void;
+  onBrowseConfiguration: () => void;
+  onCreateProfile: (controller: main.Interop_GenericController | null) => void;
+  onSaveProfile: (
+    profile: ProfileInfo,
+    controller: main.Interop_GenericController | null,
+  ) => void;
+  onEditProfile: (profile: ProfileInfo) => void;
+  onDeleteProfileForController: (profile: ProfileInfo) => void;
+};
+
+export function MainTabProfileSelector({
+  form,
+  controllers,
+  profiles,
+  onReloadConfiguration,
+  onBrowseConfiguration,
+  onCreateProfile,
+  onSaveProfile,
+  onEditProfile,
+  onDeleteProfileForController,
+}: Props) {
+  const { control, setValue } = form;
+  const selectedProfiles = useWatch({
+    name: "profiles",
+    control,
+  });
+  const selectedValue = useMemo(() => {
+    const selectedControllerProfiles = controllers.map(
+      (controller) => selectedProfiles?.[controller.GUID],
+    );
+    const hasMixedValues = selectedControllerProfiles.every(
+      (profile, index, list) =>
+        index === 0 || list[index - 1]?.Id !== profile?.Id,
+    );
+    if (hasMixedValues) return "mixed";
+    for (const profile of selectedControllerProfiles) {
+      if (profile) return profile;
+    }
+    return undefined;
+  }, [controllers, selectedProfiles]);
+  const supportedProfiles = useMemo(
+    () => profiles?.filter((profile) => !profile.UsbID),
+    [profiles],
+  );
+  const unsupportedProfiles = useMemo(
+    () => profiles?.filter((profile) => !!profile.UsbID),
+    [profiles],
+  );
+
+  const selectProfileForAll = (profile: ProfileInfo | null) => {
+    for (const controller of controllers) {
+      form.setValue(`profiles.${controller.GUID}`, profile ?? undefined);
+    }
+  };
+
+  return (
+    <fieldset className="fieldset w-full">
+      <label htmlFor="select-profile" className="fieldset-legend">
+        Select profile for all connected controllers
+      </label>
+
+      <div className="flex flex-row gap-2 items-center">
+        <div className="grow dropdown dropdown-start">
+          <button
+            id="select-profile"
+            tabIndex={0}
+            role="button"
+            className="select w-full"
+          >
+            {selectedValue
+              ? selectedValue === "mixed"
+                ? "Mixed"
+                : selectedValue.Name
+              : "Auto-detect"}
+          </button>
+          <div className="dropdown-content shadow-sm max-h-[50dvh] overflow-auto w-full">
+            <ul className="menu w-full bg-base-300 rounded-box p-2">
+              <li key="auto-detect">
+                <button
+                  className="grid grid-cols-1 grid-flow-row auto-rows-max gap-2"
+                  onClick={() => {
+                    selectProfileForAll(null);
+                    setTimeout(unfocus, 0);
+                  }}
+                >
+                  <div>Auto-detect</div>
+                </button>
+              </li>
+              {supportedProfiles.map((profile) => (
+                <ProfileSelectionListItem
+                  key={profile.Id}
+                  profile={profile}
+                  onSelect={(profile) => selectProfileForAll(profile)}
+                />
+              ))}
+              {unsupportedProfiles.map((profile) => (
+                <ProfileSelectionListItem
+                  key={profile.Id}
+                  profile={profile}
+                  disabled="Does not support all controllers"
+                />
+              ))}
+            </ul>
+          </div>
+        </div>
+
+        <ProfileSelectionMoreMenu
+          controller={null}
+          profile={selectedValue === "mixed" ? null : (selectedValue ?? null)}
+          onReloadConfiguration={onReloadConfiguration}
+          onBrowseConfiguration={onBrowseConfiguration}
+          onCreateProfile={onCreateProfile}
+          onSavePofile={onSaveProfile}
+          onEditProfile={onEditProfile}
+          onDeleteProfile={onDeleteProfileForController}
+        />
+      </div>
+    </fieldset>
+  );
+}