+++ go-app/frontend/src/utils/unfocus.ts
@@ -1,0 +1,5 @@
+export const unfocus = () => {
+  if (document.activeElement && document.activeElement instanceof HTMLElement) {
+    document.activeElement.blur();
+  }
+};