+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 1.3.6
 - Fix bug when saving calibration files
+- Add way to select profile for all controllers
 
 ## 1.3.5
 - Update webkit to 4.1 for debian/ubuntu